# src/main/resources/Testing-1.0.0-swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/JULIANAHACHMANN_1/Testing/1.0.0
  - description: API Server - Version 1.0.0
    url: https://base_url/v1
info:  
  version: "1.0.0"
  title: Estagios
  description: Uma API para conectar empresas e  instituicoes de ensino em torno de um unico objetivo - vagas de estagio!
paths:
  /me:
    get: 
      tags:
        - Perfil
      operationId: getPerfil
      summary: Meu Perfil
      description: Retorna suas informações de perfil
      responses:
        '200':
          $ref: '#/components/responses/MeuPerfil'
        '400': 
          $ref: '#/components/responses/BadRequest'
        '401': 
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':  
          $ref: '#/components/responses/TooManyRequests'
        default: 
          $ref: '#/components/responses/UnexpectedError'
      x-swagger-router-controller: ZWave

    post:
      tags:
        - Perfil
      operationId: createPerfil
      summary: Cadastrar 
      #description: ADD     
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Perfil'
      responses:
        '201':
          $ref: '#/components/responses/MeuPerfil'
        '400': 
          $ref: '#/components/responses/BadRequest'
        '401': 
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':  
          $ref: '#/components/responses/TooManyRequests'
        default: 
          $ref: '#/components/responses/UnexpectedError'
      x-swagger-router-controller: ZWave

    put:
      tags:
        - Perfil
      operationId: editPerfil
      summary: Editar perfil
      description: Editar informações do seu perfil
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Perfil'
      responses: 
        '201':
          $ref: '#/components/responses/MeuPerfil'
        '400': 
          $ref: '#/components/responses/BadRequest'
        '401': 
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':  
          $ref: '#/components/responses/TooManyRequests'
        default: 
          $ref: '#/components/responses/UnexpectedError'
      x-swagger-router-controller: ZWave

    delete: 
      tags:
        - Perfil
      operationId: deletePerfil
      summary: Deletar perfil
      description: Deletar seu perfil
      responses: 
        '204': 
          description: Perfil deletado com sucesso
        '400': 
          $ref: '#/components/responses/BadRequest'
        '401': 
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':  
          $ref: '#/components/responses/TooManyRequests'
        default: 
          $ref: '#/components/responses/UnexpectedError' 
      x-swagger-router-controller: ZWave    
...
security:
  - accountId: []
components:
  securitySchemes:
    accountId:
      type: apiKey
      in: header
      name: X-ACCOUNT
      description: All requests must include the `X-ACCOUNT` header containing your account ID.


  # ------------------------------------------------
  # PARAMETERS
...
  responses:
    MeuPerfil: 
      description: Seus dados de perfil
      content: 
        application/json: 
          schema:
            $ref: '#/components/schemas/Perfil'  
...
    BadRequest:
      description: Bad Request
      content: 
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Unauthorized: 
      description: Unauthorized
      content: 
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Forbidden: 
      description: Forbidden
      content: 
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    TooManyRequests:
      description: Too Many Requests
      content: 
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    UnexpectedError: 
      description: Unexpected Error
      content: 
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage' 


  # ----------------------------------------
  # SCHEMAS
  schemas:
...
    Endereco: 
      type: object
      properties:
        endereco: 
          type: string
          example: 'Rua João de Deus, 555'
        cidade: 
          type: string
          example: 'Florianópolis'
        estado:
          type: string
          minLength: 2
          maxLength: 2
          example: 'SC'
        pais: 
          type: string
          example: 'Brasil'
      required: 
        - cidade
        - estado
      description: "Ao menos cidade e estado devem ser fornecidos."
    Contato:
      type: object
      properties: 
        email: 
          type: string
          format: email
          example: 'contato@ies.gov.br'
        telefone:
          type: string
          example: '48 3232-2121'
      required:
        - email
      description: 'Ao menos o email deve ser fornecido'
    Perfil: 
      type: object
      properties: 
        nome: 
          type: string
          example: NoBanks
        endereco: 
          $ref: '#/components/schemas/Endereco'
        contato_geral: 
          description: 'Contato geral da instituição/ empresa'
          allOf: 
            - $ref: '#/components/schemas/Contato'
        contato_candidaturas:
          description: 'Contato para candidaturas a vagas em geral. É possível inserir contatos de candidaturas específicos quando for enviar os dados da vaga'
          allOf: 
            - $ref: '#/components/schemas/Contato'
        areas: 
          type: array
          description: 'Áreas de atuação da instituição ou empresa'
          items: 
            type: string
            example: ['fintech', 'tecnologia']
        categoria: 
          type: string
          enum: 
            - ies
            - empresa
          default: empresa
          description: "Sua categoria: Empresa ou IE  (Instituição de Ensino)"
        moderacao: 
          type: boolean
          description: Deseja ativar funções de moderação (receber denúncias de vagas, bloquear vagas ou empresas)
        notificacoes: 
          type: boolean
          description: Deseja receber notificações do sistema, tais como avisos de vagas que irão expirar, lembretes para inserir vagas
      required: 
        - nome
        - endereco
        - contato_geral
        - categoria
        - moderacao
        - notificacoes
...
    ErrorMessage:
      type: object
      properties: 
        code: 
          type: string
          minLength: 3
          maxLength: 3
          example: '400'
          description: 'HTTP Response Code'
        message: 
          type: string
          example: 'Resource could not be found'
          description: 'Error message'
      required:
        - code
        - message